refactor: refactor the repository to domain, the postman collection and the readme.
