+++ Backend/cmd/migrate/migrate.go
@@ -4,7 +4,7 @@ package main
 
 import (
 	"github.com/Game-as-a-Service/The-Message/config"
-	domain "github.com/Game-as-a-Service/The-Message/service/repository"
+	"github.com/Game-as-a-Service/The-Message/domain"
 	repository "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
 	_ "github.com/joho/godotenv/autoload"
 	"gorm.io/gorm"
