+++ Backend/domain/playerRepository.go
@@ -1,4 +1,4 @@
-package repository
+package domain
 
 import (
 	"context"
