+++ Backend/tests/e2e/game_api_test.go
@@ -4,15 +4,15 @@ import (
 	"bytes"
 	"context"
 	"encoding/json"
-	"github.com/Game-as-a-Service/The-Message/config"
-	handler "github.com/Game-as-a-Service/The-Message/service/delivery/http/v1"
-	"github.com/Game-as-a-Service/The-Message/service/repository"
-	"github.com/Game-as-a-Service/The-Message/service/service"
 	"log"
 	"net/http"
 	"net/http/httptest"
 	"os"
 	"testing"
+
+	"github.com/Game-as-a-Service/The-Message/config"
+	"github.com/Game-as-a-Service/The-Message/domain"
+	handler "github.com/Game-as-a-Service/The-Message/service/delivery/http/v1"
 
 	mysqlRepo "github.com/Game-as-a-Service/The-Message/service/repository/mysql"
 	"github.com/gin-gonic/gin"
@@ -20,7 +20,7 @@ import (
 )
 
 var serverURL string
-var gameRepo repository.GameRepository
+var gameRepo domain.GameRepository
 
 func TestMain(m *testing.M) {
 	testDB := config.InitTestDB()
@@ -29,9 +29,9 @@ func TestMain(m *testing.M) {
 
 	gameRepo = mysqlRepo.NewGameRepository(testDB)
 	playerRepo := mysqlRepo.NewPlayerRepository(testDB)
-	gameServ := service.NewGameService(gameRepo, playerRepo)
+	// gameServ := service.NewGameService(gameRepo, playerRepo)
 
-	handler.NewGameHandler(engine, gameServ)
+	handler.NewGameHandler(engine, gameRepo, playerRepo)
 
 	server := httptest.NewServer(engine)
 	serverURL = server.URL
